+++ kitty/.config/eww/scripts/change-theme.sh
@@ -6,6 +6,7 @@ ICON=""
 DOTFILES="$HOME/.dotfiles"
 THEMES="$DOTFILES/swayfx-themes"
 FOUND=false
+#DUNST_CONFIG="$HOME/.config/dunst/dunstrc"
 EWW_CONFIG="$HOME/.config/eww"
 
 die() {
@@ -24,6 +25,8 @@ uninstall_theme() {
 install_theme() {
     stow -d "$THEMES" "$1" -t "$HOME"
 }
+
+#[ -n "$DOTFILES" ] || die "No var \$DOTFILES found in your shell"
 
 [ -d "$DOTFILES" ] || die "No path $DOTFILES exist"
 [ -n "$1" ] || die "need as parameter a theme-name"
@@ -52,11 +55,17 @@ install_theme "$SELECTED"
 
 notify "$SELECTED installed, reloading...."
 
+#eww close changetheme
+#eww active-windows | grep changetheme$ | xargs eww close
+
 if eww ping -c "$EWW_CONFIG" > /dev/null; then
     echo "reloading eww..."
     eww reload -c "$EWW_CONFIG" &
     wait
 fi
+
+#~/.config/eww/scripts/start.sh &
+#wait
 
 # Reload dunst
 dunstctl reload "$HOME/.config/dunst/dunstrc" 2> /dev/null || true
@@ -70,7 +79,7 @@ tmux source-file ~/.tmux.conf 2> /dev/null || true
 # Neovim, complicated to reload plugin with lazy.nvim
 #pkill -SIGUSR1 nvim
 
-# Reload or kill...
-~/bin/reload-terminal.sh
+# Reload kitty
+pkill -SIGUSR1 kitty
 
 exit 0